renamed scenarios to more appropriate scenario inputs
